Demonstrate the use of the newer task with the copy task
The `newer` task works by comparing timestamps of source and destination files (when a task generates destination files).  For example, `newer:imagemin` will only work with source images that are newer than corresponding destination images.  It is important that two tasks are not responsible for generating the same destination file.  For example, if the `imagemin` task is responsible for generating destination images, the `copy` task should not create those same destination files.  In general, you want your tasks configured with distinct sets of files if you are going to be comparing timestamps.